Update env, separate extension checking logic
